+++ src/VlasovSolver/schemes/godunov.jl
@@ -19,7 +19,7 @@ end
 end
 
 function advect!(dest, src, g::Godunov, c, ws)
-    _validate(dest, src, g, ws)
+    _validate(dest, src, g, c, ws)
     n = length(dest)
     r, l = g.reconstruction, g.limiter
     if c > 0